refactor: Move WorldEnvelopeService to package root.
